Move State to a separate file
